# release-23.1: sql: rationalize the handling of comments and improve performance