--- a/pkg/sql/comprules/rules.go
+++ b/pkg/sql/comprules/rules.go
@@ -18,7 +18,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/compengine"
 	"github.com/cockroachdb/cockroach/pkg/sql/lexbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/scanner"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/catconstants"
 )
 
 // GetCompMethods exposes the completion heuristics defined in this
@@ -275,13 +274,13 @@ func completeObjectInCurrentDatabase(
            FROM pg_catalog.pg_class c
            JOIN pg_catalog.pg_namespace n
                 ON c.relnamespace = n.oid AND n.nspname %s
-LEFT OUTER JOIN crdb_internal.kv_catalog_comments d
-                ON c.oid = d.objoid AND d.classoid = %d
+LEFT OUTER JOIN pg_catalog.pg_description d
+                ON c.oid = d.objoid AND d.classoid = 'pg_catalog.pg_class'::REGCLASS::OID
           WHERE c.reltype != 0
             AND left(relname, length($1:::STRING)) = $1::STRING
             AND (nspname != 'pg_catalog' OR $4:::BOOL OR left($1:::STRING, 3) = 'pg_')
 `
-	query := fmt.Sprintf(queryT, schema, catconstants.PgCatalogClassTableID)
+	query := fmt.Sprintf(queryT, schema)
 	iter, err := c.Query(ctx, query, prefix, start, end, hasSchemaPrefix)
 	return iter, err
 }
@@ -310,19 +309,18 @@ func completeSchemaInCurrentDatabase(
 	}
 
 	c.Trace("completing for %q (%d,%d)", prefix, start, end)
-	const queryT = `
+	const query = `
          SELECT n.nspname AS completion,
                 'schema' AS category,
                 substr(COALESCE(d.description, ''), e'[^\n]{0,80}') as description,
                 $2:::INT AS start,
                 $3:::INT AS end
            FROM pg_catalog.pg_namespace n
-LEFT OUTER JOIN crdb_internal.kv_catalog_comments d
-                ON n.oid = d.objoid AND d.classoid = %d
+LEFT OUTER JOIN pg_catalog.pg_description d
+                ON n.oid = d.objoid AND d.classoid = 'pg_catalog.pg_namespace'::REGCLASS::OID
  WHERE left(nspname, length($1:::STRING)) = $1::STRING
 ORDER BY 1,3,4,5
 `
-	query := fmt.Sprintf(queryT, catconstants.PgCatalogNamespaceTableID)
 	iter, err := c.Query(ctx, query, prefix, start, end)
 	return iter, err
 }
@@ -425,7 +423,7 @@ func completeObjectInOtherDatabase(
 	}
 
 	c.Trace("completing for %q (%d,%d), schema: %q, db: %q", prefix, start, end, schema, dbname)
-	const queryT = `
+	const query = `
 WITH t AS (
 SELECT name, table_id
   FROM "".crdb_internal.tables
@@ -439,10 +437,9 @@ SELECT name AS completion,
        $2:::INT AS start,
        $3:::INT AS end
   FROM t
-LEFT OUTER JOIN "".crdb_internal.kv_catalog_comments cc
-    ON t.table_id = cc.objoid AND cc.classoid = %d
+LEFT OUTER JOIN "".pg_catalog.pg_description cc
+    ON t.table_id = cc.objoid AND cc.classoid = 'pg_catalog.pg_class'::REGCLASS::OID
 `
-	query := fmt.Sprintf(queryT, catconstants.PgCatalogClassTableID)
 	iter, err := c.Query(ctx, query, prefix, start, end, dbname, schema)
 	return iter, err
 }